+++ rtp.h
@@ -19,7 +19,7 @@ void rtp_shutdown(void);
 void rtp_request_resend(seq_t first, uint32_t count);
 void rtp_request_client_pause(void); // ask the client to pause
 
-void get_reference_timestamp_stuff(uint32_t *timestamp, uint64_t *timestamp_time, uint64_t *remote_timestamp_time);
+void get_reference_timestamp_stuff(int64_t *timestamp, uint64_t *timestamp_time, uint64_t *remote_timestamp_time);
 void clear_reference_timestamp(void);
 
 uint64_t static local_to_remote_time_jitters;
